requirements.txt
pyhton
django-3-jet
django-countries
django-unfold